clean and refactor spot detection
