+++ bigfish/detection/tests/test_utils.py
@@ -1,0 +1,12 @@
+# -*- coding: utf-8 -*-
+# Author: Arthur Imbert <arthur.imbert.pro@gmail.com>
+# License: BSD 3 clause
+
+"""
+Unitary tests for bigfish.stack.utils submodule.
+"""
+
+# TODO test bigfish.utils.get_sigma
+# TODO test bigfish.utils.get_radius
+# TODO test bigfish.detection.from_threshold_to_snr
+# TODO test bigfish.detection.compute_snr
